feat: log to `tracing` stream on `rosout!`
makes it more obvious that we're doing something when debugging. this isn't strictly related to the lights node, but it's what i tried first.

also, note that we're now using a patch on the `ros2-client` crate. let's try to minimize changes to be sure about our codebase... at least until we have lots of tests + some simulations :)